Only attempt to set up shell integration if running on Windows 7 or higher
The shell_integration module now checks the Windows build version.  If a build earlier than 6.1 (Windows 7) is detected, ShellIntegrationNotSupportedError is raised.  On application startup, this exception is caught, and an error dialog is displayed.  A flag is then set in the application config so the user doesn't see the dialog every time they run Codex.