Only detect DRM removal errors when converting
Previously, if a DRM removal error was detected during the conversion process, Codex would skip the file and not even attempt to convert it.  This resulted in some false positives, especially for PDFs when Adobe Digital Editions wasn't installed or configured.  This detection is now skipped unless the user has explicitly asked Codex to only remove the DRM, as the ebook-convert command will fail for protected books anyway.

Fixes #1