Begin work on documentation
Readme files, written in Markdown, now live in docs/<lang>.  These files should be saved as UTF-8.  The docs directory also contains an HTML template and stylesheet which are used to construct the final HTML versions of the readme that are distributed with the application.  python-markdown is now a required dependency because of this.  HTML files are generated when building the application executable and installer.
The application version is now 2.0 in preparation for a public release.  The English readme has yet to be written, but a Help menu has been added to the main window to allow opening of the documentation or About dialog.
Also, some old files have been removed, mostly development notes that are no longer required.
Other changes, which should've been added as separate commits but weren't: If running from source, Python tracebacks are printed to sys.stdout as normal instead of being logged.  Debug logging has been disabled.